+++ js/main.js
@@ -1,9 +1,9 @@
 import {getUsers} from './data.js';
 import {renderPicturePreviews} from './picture.js';
 import {addOpenHandler} from './bigpicture-viewer.js';
-import {newImageUploader} from './file-uploader.js';
+import {initImageUploader} from './image-uploader.js';
 
 const picturesData = getUsers();
 renderPicturePreviews(picturesData);
 addOpenHandler(picturesData);
-newImageUploader();
+initImageUploader();
